feat : 로그아웃 성공 핸들러 추가 #83
