+++ server/src/main/java/run/ward/mmz/handler/auth/CustomLoginSuccessHandler.java
@@ -1,0 +1,26 @@
+package run.ward.mmz.handler.auth;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.stereotype.Service;
+
+import java.io.IOException;
+
+@Service
+@RequiredArgsConstructor
+public class CustomLoginSuccessHandler implements AuthenticationSuccessHandler {
+
+
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
+        String uri = request.getHeader("Referer");
+        if (uri != null && !uri.contains("/api/v1/auth/login")) {
+            request.getSession().setAttribute("prevPage", uri);
+        }
+    }
+}
