+++ server/src/main/java/run/ward/mmz/handler/auth/CustomLogoutSuccessHandler.java
@@ -1,0 +1,20 @@
+package run.ward.mmz.handler.auth;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public class CustomLogoutSuccessHandler implements LogoutSuccessHandler {
+
+    @Override
+    public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
+        String uri = request.getHeader("Referer");
+        if (uri != null && !uri.contains("/api/v1/auth/login")) {
+            request.getSession().setAttribute("prevPage", uri);
+        }
+    }
+}
