--- server/src/main/java/run/ward/mmz/handler/auth/LoginSuccessHandler.java
@@ -1,50 +1,0 @@
-package run.ward.mmz.handler.auth;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.security.web.savedrequest.HttpSessionRequestCache;
-import org.springframework.security.web.savedrequest.RequestCache;
-import org.springframework.security.web.savedrequest.SavedRequest;
-import org.springframework.stereotype.Service;
-import run.ward.mmz.domain.account.Account;
-import run.ward.mmz.handler.exception.CustomException;
-import run.ward.mmz.handler.exception.ExceptionCode;
-import run.ward.mmz.repository.AccountRepository;
-
-import java.io.IOException;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-@Service
-@RequiredArgsConstructor
-public class LoginSuccessHandler implements AuthenticationSuccessHandler {
-
-    private final AccountRepository accountRepository;
-
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request,
-                                        HttpServletResponse response,
-                                        Authentication authentication) throws IOException {
-
-        RequestCache requestCache = new HttpSessionRequestCache();
-        SavedRequest savedRequest = requestCache.getRequest(request, response);
-
-
-        String url = "/api/v1/auth/login-success";
-
-        Account user = accountRepository.findByEmail(authentication.getName()).orElseThrow(
-                () -> new CustomException(ExceptionCode.USER_NOT_FOUND)
-        );
-
-        if(user.isNew()){
-            url = "/api/v1/auth/signup/user-info";
-        }
-
-        response.sendRedirect(url);
-    }
-
-}
