+++ server/src/main/java/run/ward/mmz/service/account/impl/AccountServiceImpl.java
@@ -14,7 +14,6 @@ import run.ward.mmz.handler.exception.ExceptionCode;
 import run.ward.mmz.repository.AccountRepository;
 import run.ward.mmz.service.account.AccountService;
 
-import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
